+++ lib/runners/processor/coffeelint.rb
@@ -30,7 +30,6 @@ module Runners
         return Results::Failure.new(guid: guid, message: exn.message, analyzer: nil)
       end
 
-      analyzer # Must initialize after installation
       yield
     end