+++ lib/runners/processor/querly.rb
@@ -30,7 +30,6 @@ module Runners
 
     def setup
       ret = install_gems default_gem_specs, optionals: OPTIONAL_GEMS, constraints: CONSTRAINTS do
-        analyzer
         yield
       end