+++ lib/runners/processor/rails_best_practices.rb
@@ -47,7 +47,6 @@ module Runners
 
       prepare_config
       install_gems default_gem_specs, optionals: OPTIONAL_GEMS, constraints: CONSTRAINTS do
-        analyzer
         yield
       end
     rescue InstallGemsFailure => exn