+++ lib/runners/processor/tslint.rb
@@ -40,7 +40,6 @@ module Runners
         return Results::Failure.new(guid: guid, message: exn.message, analyzer: nil)
       end
 
-      analyzer # Must initialize after installation
       yield
     end